Validations, rules, image (upload and validations), regex and add Dapper

+++ src/Paulino.Motorbike.Domain/Motorbike/Handlers/DeleteByIdMotorbikeHandler.cs
@@ -2,6 +2,9 @@
 using Microsoft.EntityFrameworkCore;
 using Paulino.Motorbike.Domain.Motorbike.Requests;
 using Paulino.Motorbike.Infra.CrossCutting.Exceptions;
+using Paulino.Motorbike.Infra.Data.Dapper.Base;
+using Paulino.Motorbike.Infra.Data.Dapper.Dtos;
+using Paulino.Motorbike.Infra.Data.Dapper.Queries;
 using Paulino.Motorbike.Infra.Data.EF;
 
 namespace Paulino.Motorbike.Domain.Motorbike.Handlers
@@ -9,10 +12,12 @@ namespace Paulino.Motorbike.Domain.Motorbike.Handlers
     public class DeleteByIdMotorbikeHandler : IRequestHandler<DeleteByIdMotorbikeRequest, Unit>
     {
         private readonly IApplicationDbContext _dbContext;
+        private readonly IDapperRepository _dapper;
 
-        public DeleteByIdMotorbikeHandler(IApplicationDbContext dbContext)
+        public DeleteByIdMotorbikeHandler(IApplicationDbContext dbContext, IDapperRepository dapper)
         {
             _dbContext = dbContext;
+            _dapper = dapper;
         }
 
         public async Task<Unit> Handle(DeleteByIdMotorbikeRequest request, CancellationToken cancellationToken)
@@ -21,6 +26,11 @@ namespace Paulino.Motorbike.Domain.Motorbike.Handlers
 
             if (motorbike == null)
                 throw new BadRequestException("Dados inválidos");
+
+            var activeRentals = await _dapper.QueryAsync<GetActiveRentalsByMotorbikeDapperQueryDto>(new GetActiveRentalsByMotorbikeDapperQuery(request.Id));
+
+            if ((activeRentals?.Count() ?? 0) > 0)
+                throw new BadRequestException("Tem locação ativa");
 
             _dbContext.Motorbike.Remove(motorbike);
             await _dbContext.SaveChangesAsync();
